fabtests/common: fix transfer_size op parsing
transfer_size is a size_t, change atoi to atol to
support large sizes

+++ fabtests/common/shared.c
@@ -2971,7 +2971,7 @@ void ft_parsecsopts(int op, char *optarg, struct ft_opts *opts)
 			opts->sizes_enabled = FT_ENABLE_ALL;
 		} else {
 			opts->options |= FT_OPT_SIZE;
-			opts->transfer_size = atoi(optarg);
+			opts->transfer_size = atol(optarg);
 		}
 		break;
 	case 'm':
